Update to template.xlsm
To support the later versions of MS Excel the macro for importing images
has been altered.
Normally the image was inserted and saved, but with the Excel file, but
with the later version, the image needs to be inserted in a specific
way. If not, the image is only linked to and not saved with the
document.